Refactoring and more tests.

+++ src/Toe.SPIRV/Instructions/OpCopyMemory.cs
@@ -38,7 +38,7 @@ namespace Toe.SPIRV.Instructions
             uint wordCount = 0;
             wordCount += Target.GetWordCount();
             wordCount += Source.GetWordCount();
-            wordCount += MemoryAccess?.GetWordCount() ?? (uint)0;
+            wordCount += MemoryAccess?.GetWordCount() ?? 0u;
             return wordCount;
         }
