+++ src/Toe.SPIRV/Instructions/OpCopyMemorySized.cs
@@ -43,7 +43,7 @@ namespace Toe.SPIRV.Instructions
             wordCount += Target.GetWordCount();
             wordCount += Source.GetWordCount();
             wordCount += Size.GetWordCount();
-            wordCount += MemoryAccess?.GetWordCount() ?? (uint)0;
+            wordCount += MemoryAccess?.GetWordCount() ?? 0u;
             return wordCount;
         }
