+++ src/Toe.SPIRV/Instructions/OpImageWrite.cs
@@ -43,7 +43,7 @@ namespace Toe.SPIRV.Instructions
             wordCount += Image.GetWordCount();
             wordCount += Coordinate.GetWordCount();
             wordCount += Texel.GetWordCount();
-            wordCount += ImageOperands?.GetWordCount() ?? (uint)0;
+            wordCount += ImageOperands?.GetWordCount() ?? 0u;
             return wordCount;
         }
